Move some classes to upper package
