+++ storio/src/main/java/com/pushtorefresh/storio/contentprovider/operation/put/PreparedPut.java
@@ -3,7 +3,7 @@ package com.pushtorefresh.storio.contentprovider.operation.put;
 import android.support.annotation.NonNull;
 
 import com.pushtorefresh.storio.contentprovider.StorIOContentProvider;
-import com.pushtorefresh.storio.db.operation.PreparedOperation;
+import com.pushtorefresh.storio.operation.PreparedOperation;
 
 public abstract class PreparedPut<T> implements PreparedOperation<T> {
