+++ storio/src/main/java/com/pushtorefresh/storio/db/operation/put/PreparedPut.java
@@ -4,7 +4,7 @@ import android.content.ContentValues;
 import android.support.annotation.NonNull;
 
 import com.pushtorefresh.storio.db.StorIODb;
-import com.pushtorefresh.storio.db.operation.PreparedOperation;
+import com.pushtorefresh.storio.operation.PreparedOperation;
 
 import java.util.Arrays;
