+++ storio/src/main/java/com/pushtorefresh/storio/db/operation/put/PreparedPutIterableContentValues.java
@@ -4,7 +4,7 @@ import android.content.ContentValues;
 import android.support.annotation.NonNull;
 
 import com.pushtorefresh.storio.db.StorIODb;
-import com.pushtorefresh.storio.db.operation.Changes;
+import com.pushtorefresh.storio.db.Changes;
 import com.pushtorefresh.storio.util.EnvironmentUtil;
 
 import java.util.HashMap;
