+++ storio/src/main/java/com/pushtorefresh/storio/db/operation/put/PreparedPutWithObject.java
@@ -4,9 +4,9 @@ import android.content.ContentValues;
 import android.support.annotation.NonNull;
 
 import com.pushtorefresh.storio.db.StorIODb;
-import com.pushtorefresh.storio.db.operation.Changes;
-import com.pushtorefresh.storio.db.operation.MapFunc;
-import com.pushtorefresh.storio.db.operation.PreparedOperation;
+import com.pushtorefresh.storio.db.Changes;
+import com.pushtorefresh.storio.operation.MapFunc;
+import com.pushtorefresh.storio.operation.PreparedOperation;
 import com.pushtorefresh.storio.util.EnvironmentUtil;
 
 import rx.Observable;
