+++ storio/src/test/java/com/pushtorefresh/storio/db/unit_test/operation/PreparedExecSqlTest.java
@@ -1,7 +1,7 @@
 package com.pushtorefresh.storio.db.unit_test.operation;
 
 import com.pushtorefresh.storio.db.StorIODb;
-import com.pushtorefresh.storio.db.operation.Changes;
+import com.pushtorefresh.storio.db.Changes;
 import com.pushtorefresh.storio.db.operation.exec_sql.PreparedExecSql;
 import com.pushtorefresh.storio.db.query.RawQuery;
